feature(main): finishing project

+++ ToDoList/Views/ListView.swift
@@ -13,22 +13,29 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List{
-            // there's no need for an ID because we already made one at our item model
-            ForEach(listViewModel.items)  { item in
-                ListRowView(item: item)
-                // what's going to happen when you touch an item
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            // If it's empty, show message
+            if listViewModel.items.isEmpty {
+               NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List{
+                    // there's no need for an ID because we already made one at our item model
+                    ForEach(listViewModel.items)  { item in
+                        ListRowView(item: item)
+                        // what's going to happen when you touch an item
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+//                .listStyle(InsetListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-            
         }
-        .listStyle(PlainListStyle())
         // customizing navigation and navItems
         .navigationTitle("To do list")
         .navigationBarItems(
